Refactor createAnimatableComponent to support custom animations
